test: move WatchFolderServiceTests to a dedicated file
